Changed the separator from Sheffer stroke | to exclamation point !
Looked into encodeURIcomponent and found tha it leaves ! alone.  It also leaved ~ along. These can be in file name, too. So, that fixes problems. One problem is that Chromium based browsers will encode the file name in the browser and fail to open the file if you are looking at something on disk.