Policy on formats - don't wait around
Delayed loading of formats will be acceptable in the browser version of CWID. But, for node.js, the supported formats and bases will be loaded through a require.

REQUIREMENT: A tool will be required to update the supported formats... It must be accessible to users above the node_module level.